# docs/openapi-person-api.json
{"swagger": "2.0", "basePath": "/", "paths": {"/api/persons": {"post": {"responses": {"202": {"description": "Person has been successfully created"}, "500": {"description": "Internal server error"}}, "description": "Creates a new Person", "operationId": "post_persons_resource", "parameters": [{"name": "payload", "required": true, "in": "body", "schema": {"$ref": "#/definitions/Person"}}], "tags": ["UdaConnect - Person API"]}, "get": {"responses": {"200": {"description": "Successfully retrieved all people", "schema": {"type": "array", "items": {"$ref": "#/definitions/Person"}}}}, "description": "Retrieve all people", "operationId": "get_persons_resource", "tags": ["UdaConnect - Person API"]}}, "/api/persons/{person_id}": {"get": {"responses": {"200": {"description": "Person was found", "schema": {"$ref": "#/definitions/Person"}}, "404": {"description": "Person not found"}, "500": {"description": "Internal server error"}}, "description": "Search for a given person by id", "operationId": "get_person_resource", "parameters": [{"in": "path", "description": "Id(required) of the person to search", "name": "person_id", "required": true, "type": "string"}], "tags": ["UdaConnect - Person API"]}}, "/api/persons/{person_id}/connection": {"parameters": [{"in": "query", "description": "Proximity in meters", "name": "distance", "type": "string"}, {"in": "query", "description": "end of date range", "name": "end_date", "type": "string"}, {"in": "query", "description": "start of date range", "name": "start_date", "type": "string"}, {"name": "person_id", "in": "path", "required": true, "type": "string"}], "get": {"responses": {"200": {"description": "Succesfully found connection(s)", "schema": {"type": "array", "items": {"$ref": "#/definitions/Connection"}}}}, "description": "Searches for connection(s) of a given person", "operationId": "get_connection_data_resource", "tags": ["UdaConnect - Person API"]}}}, "info": {"title": "UdaConnect - Person API", "version": "1.0.0"}, "produces": ["application/json"], "consumes": ["application/json"], "tags": [{"name": "UdaConnect - Person API", "description": "Endpoints to provide person data"}], "definitions": {"Person": {"properties": {"id": {"type": "integer"}, "first_name": {"type": "string"}, "last_name": {"type": "string"}, "company_name": {"type": "string"}}, "type": "object"}, "Connection": {"properties": {"location": {"$ref": "#/definitions/Location"}, "person": {"$ref": "#/definitions/Person"}}, "type": "object"}, "Location": {"properties": {"id": {"type": "integer"}, "person_id": {"type": "integer"}, "longitude": {"type": "string"}, "latitude": {"type": "string"}, "creation_time": {"type": "string", "format": "date-time"}}, "type": "object"}}, "responses": {"ParseError": {"description": "When a mask can't be parsed"}, "MaskError": {"description": "When any error occurs on mask"}}}